requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
azure-cosmos